--- ui/src/app/base/components/Pager/index.js
@@ -1,1 +1,0 @@
-export { default } from "./Pager";
