--- ui/src/app/base/components/PagerButton/index.js
@@ -1,1 +1,0 @@
-export { default } from "./PagerButton";
