--- ui/src/app/base/components/PagerItem/index.js
@@ -1,1 +1,0 @@
-export { default } from "./PagerItem";
