+++ ui/src/app/base/components/PaginationButton/index.js
@@ -1,0 +1,1 @@
+export { default } from "./PaginationButton";
