+++ ui/src/app/base/components/PaginationItem/PaginationItem.test.js
@@ -1,26 +1,30 @@
 import { shallow } from "enzyme";
 import React from "react";
 
-import PagerItem from "./PagerItem";
+import PaginationItem from "./PaginationItem";
 
-describe("<PagerItem />", () => {
+describe("<PaginationItem />", () => {
   // snapshot tests
   it("renders and matches the snapshot", () => {
-    const component = shallow(<PagerItem number={1} onClick={jest.fn()} />);
+    const component = shallow(
+      <PaginationItem number={1} onClick={jest.fn()} />
+    );
 
     expect(component).toMatchSnapshot();
   });
 
   // unit tests
   it("displays the page number", () => {
-    const component = shallow(<PagerItem number={1} onClick={jest.fn()} />);
+    const component = shallow(
+      <PaginationItem number={1} onClick={jest.fn()} />
+    );
 
     expect(component.find("button").text()).toEqual("1");
   });
 
   it("sets isActive", () => {
     const component = shallow(
-      <PagerItem number={1} onClick={jest.fn()} isActive />
+      <PaginationItem number={1} onClick={jest.fn()} isActive />
     );
 
     expect(component.find("button").hasClass("is-active"));
