+++ ui/src/app/base/components/PaginationItem/index.js
@@ -1,0 +1,1 @@
+export { default } from "./PaginationItem";
